+++ src/pages/Devtools/devtools.css
@@ -1,0 +1,6 @@
+#root {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
