+++ src/pages/Register/index.tsx
@@ -1,0 +1,11 @@
+/** @jsxImportSource @emotion/react */
+import { useNavigate } from 'react-router-dom'
+// import { registerStyles } from './styles'
+
+const Register = () => {
+  //   const navigate = useNavigate()
+
+  return <div>Register</div>
+}
+
+export default Register
